requirements.txt
asgiref==3.8.1
axios==0.4.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
django-rest-framework-docs==0.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==3.2.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3